Changes to Player Attacks
Attack window is determined by a curve, CloseAttackWindow notify will set up the next attack in the sequence if an attack is queued (will remove this notify once I can get the code to work with the float properly).

Dodge roll is in a better state, cannot spam it and it cannot be interrupted by attacks.