Combat Overhaul
- New Camera Manager -

The new camera manager introduces four new camera setups: Exploration, Action, Aiming and Lock-on.

Exploration: Used when the character is not in combat mode.

Combat: Used when the character is in combat mode, it provides the player with a greater view of their surroundings.

Aiming: Used when the character is aiming with a ranged weapon, it positions the camera next to the character’s shoulder and rotates the character as the camera moves.

Lock-on: Used when the character has locked on to a target, the camera will focus on the target and keep it in the centre of the viewport, the player can swap between targets based on their mouse/controller input. Lock-on will automatically swap back to the previous camera position if the current target is no longer targetable.

- Controller Input HUD -

Dynamically updates based on the character’s combat mode and combat stance.

The context button’s prompt changes depending on if there is an interactable actor in range or if the character is moving fast enough to dodge instead.

The HUD will use different input icons depending on the controller being used (e.g. PS4 Dualshock, Xbox).

- Weapons -

Added bows and greatswords.

Added blocking functionality to weapons, stamina and condition are both reduced when the weapon blocks an attack.

The condition of a weapon decreases when hitting actors during an attack.

 - Attack Combos -

Weapons have both a light attack combo and a heavy attack combo.

You can alternate between light and heavy attacks during a combo.

Holding the attack button during a combo will make the character do a combo finisher.

If using the Lock-on camera, attacks will adjust their angle based on the height of the lock-on target.

Damage dealt by attacks are partially random, each attack in a combo has its own base damage value which is then added to a random number ranging between 0 and the attacking weapon’s WeaponDamageVariance variable.

 - Combat Stances -

Currently there are two combat stances: Guard and Aim.

Guard: Allows the character to block with their currently equipped weapon/s, this stance prevents the character from rotating while also reducing their movement speed and stamina regeneration speed.

Aim: Allows the character to aim with their weapon at the crosshair.

Each weapon can have an offhand and mainhand stance.

In the future, combat stances with give the player access to weapon skills that can be used instead of light and heavy attacks.

 - Combat Actions -

Currently there is only one combat action for the bow that allows the character to draw an arrow with and release it towards the crosshair, if the release of the arrow is timed correctly then the arrow will fly further and deal more damage.

Each weapon can have 3 types of combat action, one when they are not in a combat stance and one each for their offhand and mainhand combat stances.

- Stamina Usage & Exhaustion -

Attacks, combat actions, dodging and blocking all reduce stamina.

If a character completely drains their stamina, they will put into a state of exhaustion where they cannot act until their stamina has partially recovered.

- Dodge -

Has invulnerability frames during the animation to prevent the character from being harmed.

 - Update: Health Component -

Health is now divided into body parts, each storing a percentage of the component’s max HP. Each body part has an array of bones it is related to, if those bones are hit then the related body part will take damage.

Getting hit in a particular body part will reduce the health of that body part based on its damage modifier (e.g. The head body part takes 25% increased damage).

When a body part has no health, it will be labelled as “broken”, broken parts will distribute the damage they take between all remaining body parts that are still healthy.

If a body part labelled as “vital” is broken, then the character dies.

 - Other Changes -

New action binding to swap between weapon loadouts (can be done anytime).

Actors hit with an attack will play a sound cue and spawn a basic VFX depending on its physical material.